% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipRotator.R
\name{create_resource}
\alias{create_resource}
\title{Create Ressource}
\usage{
create_resource(svc)
}
\arguments{
\item{svc}{Input your initialized Gateway.}
}
\description{
Creates Resource
}
\examples{
create_resource(svc)
}
