% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipRotator.R
\name{create_rest_api}
\alias{create_rest_api}
\title{Create Rest API}
\usage{
create_rest_api(svc, url)
}
\arguments{
\item{svc}{Input your initialized Gateway.}
}
\description{
Creates Rest API
}
\examples{
create_rest_api(svc)
}
