% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipRotator.R
\name{put_integration}
\alias{put_integration}
\title{Put Integration}
\usage{
put_integration(url)
}
\arguments{
\item{url}{The target URL you want rotating requesting IPs on}
}
\description{
Creates Integration
}
\examples{
put_integration("https://www.google.com")
}
