% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipRotator.R
\name{rotated_get}
\alias{rotated_get}
\title{Rotated GET}
\usage{
rotated_get(url)
}
\arguments{
\item{url}{The URL defined in the setup of the Rest API + site path (if needed)}
}
\description{
Request URL with rotated_get to request with rotated ips
}
\examples{
rotated_get("https://www.reddit.com/r/all/")
}
