% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipRotator.R
\name{run_all}
\alias{run_all}
\title{Run All}
\usage{
run_all(access_key, secret_access_key, region, url)
}
\arguments{
\item{access_key}{Your AWS Access Key ID}

\item{secret_access_key}{Your AWS Secret Access Key ID}

\item{region}{"us-east-1", "us-east-2", "us-west-1", "us-west-2", "eu-west-1", "eu-west-2", "eu-west-3", "eu-north-1", "eu-central-1", "ca-central-1"}

\item{url}{The target URL you want rotating requesting IPs on}
}
\description{
Runs all steps to create the IP Rotating Rest-API
}
\examples{
run_all("AKIB8KIA9AO85K8IZ5HJ","nbLRi0qiMSoAGybbV2c86SEl2nhIMPNc5fagV7KQ", "eu-central-1", "https://www.google.com")
}
